[xpmp_sms] fix song selection for non-FM mode
The song table includes all 13 SMS channels (FM) even when FM is not enabled.
The most efficient solution would be to have XPMP/XPMC not incorporate any usage
of the FM channels when they are not enabled. but we can see in the ASM file output
by XPMC that this it not the case for pattern data and song table.

This is the least intrusive solution to XPMC/XPMP for now.