[xpmc] Add DEFINE compiler directive and state the feature is available in a definition
I had the desire to implement this feature for one reason. The following:
XPMC allows the creation of multiple songs compiled from one MML file. The creates a song engine for something like a video game, which I'm interested in making. An interesting case is that all the songs pick from the same pool of macros (this may be considered a current limitation of XPMCK, or just the way things are.) So, not all systems have easy-peezy support for selecting that non-1st song in a Song list, so I like to develop songs separately and then have a master.mml file which compiles them all together into one master compilation. To do this, I maintain a "common.mml" file which has all the envelopes / macros for all of the songs... THE PROBLEM lies here -- in order to develop each song individually, you must #INCLUDE "common.mml" -- but when you have master.mml including each song's mml file, it recursively includes the common.mml file!! By the use of #DEFINE .. I can add a symbol check in the same style of C header files, which ensures that it is not recursively defined its contents.

I realize in that lengthy description that an example archive demonstrating everything I've talked about is in order. I will see to including an external link in this commit message (although it is likely not to be a very long-lasting link.. Therefore I may include the archive in the repo itself)